# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Library API
  version: "1.0.0"
  description: Simple library management REST API (BSc project).
servers:
  - url: http://localhost:5000/api
paths:
...
  /register:
    post:
      summary: Register new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegisterResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
  /login:
    post:
      summary: Login and obtain tokens
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /token/refresh:
    post:
      summary: Refresh access token
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
        "401":
          $ref: "#/components/responses/Unauthorized"
  /logout:
    post:
      summary: Logout (revoke current token)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
        "401":
          $ref: "#/components/responses/Unauthorized"
  /me:
    get:
      summary: Current user claims
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  user_id: { type: integer, example: 1 }
                  role: { type: string, example: Member }
                  library_id: { type: integer, example: 1 }
        "401":
          $ref: "#/components/responses/Unauthorized"
  /me/password:
    post:
      summary: Change own password
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [old_password, new_password]
              properties:
                old_password: { type: string }
                new_password: { type: string, description: "Min 8 chars, must include letter and digit." }
      responses:
        "200":
          description: Password changed
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string, enum: [ok] }
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/ServerError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    NotFound:
      description: Not found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Conflict:
      description: Conflict
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    ServerError:
      description: Server error
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    Error:
      type: object
      properties:
        error: { type: string, example: invalid_credentials }
        message: { type: string }
        meta:
          type: object
          additionalProperties: true
    RegisterRequest:
      type: object
      required: [email, password, name, address, date_of_birth]
      properties:
        email: { type: string, format: email }
        password: { type: string, description: "Min 8 chars, include letter and digit." }
        name: { type: string }
        address: { type: string }
        date_of_birth: { type: string, example: 2000-01-01 }
    RegisterResponse:
      type: object
      properties:
        user_id: { type: integer }
        email: { type: string }
        name: { type: string }
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string }
    LoginResponse:
      type: object
      properties:
        access_token: { type: string }
        refresh_token: { type: string }
        user:
          type: object
          properties:
            user_id: { type: integer }
            email: { type: string }
            name: { type: string }
            role: { type: string }
            library_id: { type: integer }
...